split main into multiple modules
