+++ constants.js
@@ -1,0 +1,18 @@
+export const BASE_SPRITE_X_OFFSET = 10
+export const BASE_SPRITE_Y_OFFSET = 30
+export const FRAME_STAGGER = 5
+export const SFX = {
+        jump: new Howl({
+            src: './sfx/cartoon-jump-6462.mp3',
+            volume: 0.5
+        }),
+        collect: new Howl({
+            src: './sfx/coin-257878.mp3',
+            volume: 0.5
+        }),
+        spawn: new Howl({
+            src: './sfx/portal-phase-jump-6355.mp3',
+            volume: 0.6
+        })
+
+    }
