+++ index.html
@@ -34,7 +34,13 @@
         <h3>Level 2</h3>
         <div class="debug-section" id="level2Debug"></div>
     </div>
-    <script src="main.js"></script>
+    
+    <script type="module" src="player.js"></script>
+    <script type="module" src="platform.js"></script>
+    <script type="module" src="background.js"></script>
+    <script type="module" src="coin.js"></script>
+    <script type="module" src="portal.js"></script>
+    <script type="module" src="main.js"></script>
     <script src="https://cdnjs.cloudflare.com/ajax/libs/howler/2.2.4/howler.min.js" integrity="sha512-xi/RZRIF/S0hJ+yJJYuZ5yk6/8pCiRlEXZzoguSMl+vk2i3m6UjUO/WcZ11blRL/O+rnj94JRGwt/CHbc9+6EA==" crossorigin="anonymous" referrerpolicy="no-referrer"></script>
 </body>
 </html>
